misc: change `uv_ip4_addr()` `port` parameter type
Change the `port` parameter of `uv_ip4_addr()` and `uv_ip6_addr()` from
an `int` to an `unsigned short`.

Currently, the `uv_ip4_addr()` and `uv_ip6_addr()` functions cast the
`port` to a `uint16_t` internally anyway, when calling [`htons()`][1].

Making the function accept `unsigned short` instead has two benefits:
  - Very slight performance increase, since we can avoid a
    `unsigned short` -> `int` -> `unsigned short` conversion.
  - Compilers can warn users of `libuv` if their `port` parameter
    to `uv_ip4_addr()` might be truncated.

BREAKING CHANGE: This change breaks the ABI, but it does not break the
                 API (other than potentially causing some compiler
                 warnings if the user calls this function with
                 overflowing types).

[1]: https://linux.die.net/man/3/htons

+++ include/uv.h
@@ -1763,8 +1763,8 @@ UV_EXTERN int uv_fs_event_getpath(uv_fs_event_t* handle,
                                   char* buffer,
                                   size_t* size);
 
-UV_EXTERN int uv_ip4_addr(const char* ip, int port, struct sockaddr_in* addr);
-UV_EXTERN int uv_ip6_addr(const char* ip, int port, struct sockaddr_in6* addr);
+UV_EXTERN int uv_ip4_addr(const char* ip, unsigned short port, struct sockaddr_in* addr);
+UV_EXTERN int uv_ip6_addr(const char* ip, unsigned short port, struct sockaddr_in6* addr);
 
 UV_EXTERN int uv_ip4_name(const struct sockaddr_in* src, char* dst, size_t size);
 UV_EXTERN int uv_ip6_name(const struct sockaddr_in6* src, char* dst, size_t size);
